#161 Deprecated Channel interface and MultiplexedChannel class. Moved new channel classes to proper package.
