+++ io/network/src/main/java/org/mobicents/media/server/io/network/channel/AbstractNetworkChannel.java
@@ -19,7 +19,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.mobicents.media.server.io.network.channel2;
+package org.mobicents.media.server.io.network.channel;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
